Move request pkg to internal dir
