+++ internal/request/graphql/schema/examples/example.schema.gql
