+++ internal/request/graphql/schema/examples/root.schema.gql
