+++ internal/request/graphql/schema/index_parse_test.go
