+++ internal/request/graphql/schema/manager.go
@@ -13,7 +13,7 @@ package schema
 import (
 	gql "github.com/sourcenetwork/graphql-go"
 
-	schemaTypes "github.com/sourcenetwork/defradb/request/graphql/schema/types"
+	schemaTypes "github.com/sourcenetwork/defradb/internal/request/graphql/schema/types"
 )
 
 // SchemaManager creates an instanced management point
