+++ internal/request/graphql/schema/type.schema.gen.gql.template
