+++ internal/request/graphql/schema/types/base.go
