+++ tests/bench/query/planner/utils.go
@@ -21,8 +21,8 @@ import (
 	acpIdentity "github.com/sourcenetwork/defradb/internal/acp/identity"
 	"github.com/sourcenetwork/defradb/internal/core"
 	"github.com/sourcenetwork/defradb/internal/planner"
-	"github.com/sourcenetwork/defradb/request/graphql"
-	gqlSchema "github.com/sourcenetwork/defradb/request/graphql/schema"
+	"github.com/sourcenetwork/defradb/internal/request/graphql"
+	gqlSchema "github.com/sourcenetwork/defradb/internal/request/graphql/schema"
 	benchutils "github.com/sourcenetwork/defradb/tests/bench"
 	"github.com/sourcenetwork/defradb/tests/bench/fixtures"
 )
